requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1
django-crispy-forms==1.9.2
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
Pillow==7.2.0
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
